Call reset when fetching fuel logs.
This fixes an issue where requests after the first would have no effect. Backbone has the collection loaded and thought it could get away without re-fetching. While this would be nice in the future for performance reasons, we need a smart way to keep track of the request; page, offset, etc